Rename app to server
